Consolidate formBuilderUtil
